requirements.txt
python-dotenv
pyodbc
requests
python-dateutil